Changed implementation to DistributedCache
