--- Backend/Remora.Discord.Caching.Redis/Remora.Discord.Caching.Redis.csproj
@@ -1,25 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFrameworks>netstandard2.1;net5.0</TargetFrameworks>
-        <RootNamespace>Remora.Discord.Caching</RootNamespace>
-    </PropertyGroup>
-
-    <PropertyGroup>
-        <Version>2.0.0</Version>
-        <Description>
-            $(BaseDescription)
-
-            This package contains memory cache functionality for the gateway and REST API.
-        </Description>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\Remora.Discord.Caching\Remora.Discord.Caching.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <PackageReference Include="StackExchange.Redis" Version="2.2.4" />
-    </ItemGroup>
-
-</Project>
