+++ Backend/Remora.Discord.Caching/Extensions/ServiceCollectionExtensions.cs
@@ -26,6 +26,7 @@ using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Options;
 using Remora.Discord.API.Abstractions.Rest;
 using Remora.Discord.Caching.API;
+using Remora.Discord.Caching.Clients;
 using Remora.Discord.Caching.Responders;
 using Remora.Discord.Caching.Services;
 using Remora.Discord.Gateway.Extensions;
