+++ Remora.Discord.sln
@@ -77,8 +77,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Remora.Discord.Caching", "B
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Caching", "Samples\Caching\Caching.csproj", "{E8B6A214-6F37-48E7-8E10-77C1DBD7DF09}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Remora.Discord.Caching.Redis", "Backend\Remora.Discord.Caching.Redis\Remora.Discord.Caching.Redis.csproj", "{99D138E7-6AE9-4800-9D5C-8C8405F048A4}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -169,10 +167,6 @@ Global
 		{E8B6A214-6F37-48E7-8E10-77C1DBD7DF09}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E8B6A214-6F37-48E7-8E10-77C1DBD7DF09}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E8B6A214-6F37-48E7-8E10-77C1DBD7DF09}.Release|Any CPU.Build.0 = Release|Any CPU
-		{99D138E7-6AE9-4800-9D5C-8C8405F048A4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{99D138E7-6AE9-4800-9D5C-8C8405F048A4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{99D138E7-6AE9-4800-9D5C-8C8405F048A4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{99D138E7-6AE9-4800-9D5C-8C8405F048A4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{DC82899D-871E-4DC9-B61C-7619111800DE} = {26E93608-29BA-45C7-993A-AA24A49D6013}
@@ -194,6 +188,5 @@ Global
 		{48DFFD60-CD36-47C7-90E4-A82D8EB0B1A0} = {6E2D59EB-648D-4968-8A63-8D8E3813EC86}
 		{7B902015-C8E7-496C-8030-5094CCC93800} = {26E93608-29BA-45C7-993A-AA24A49D6013}
 		{E8B6A214-6F37-48E7-8E10-77C1DBD7DF09} = {D1A34922-E02F-492B-AEF5-9B34E81CE293}
-		{99D138E7-6AE9-4800-9D5C-8C8405F048A4} = {26E93608-29BA-45C7-993A-AA24A49D6013}
 	EndGlobalSection
 EndGlobal
