+++ Samples/SlashCommands/SlashCommands.csproj
@@ -7,12 +7,11 @@
     </PropertyGroup>
 
     <ItemGroup>
+        <PackageReference Include="Microsoft.Extensions.Caching.StackExchangeRedis" Version="5.0.1" />
         <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="5.0.0" />
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\..\Backend\Remora.Discord.Caching.Memory\Remora.Discord.Caching.Memory.csproj" />
-        <ProjectReference Include="..\..\Backend\Remora.Discord.Caching.Redis\Remora.Discord.Caching.Redis.csproj" />
         <ProjectReference Include="..\..\Backend\Remora.Discord.Caching\Remora.Discord.Caching.csproj" />
         <ProjectReference Include="..\..\Backend\Remora.Discord.Gateway\Remora.Discord.Gateway.csproj" />
         <ProjectReference Include="..\..\Remora.Discord.Commands\Remora.Discord.Commands.csproj" />
